Implement a cache mechanism for /movies command
Theaters keep their movies for about a week. We don't need a fresh movie
list every day or hour. Use a map to cache the result for a week using
ISO week (week of the year). Since every year has at most 53 weeks and
given the restart period of the ilberbot service, we don't need to
invalidate caches ever. This keeps the implementation straightforward.