build: add a goxc configuration file
goxc has no knowledge of Go vendoring, hence it tries to test and vet
all the dependencies for every build, and it's starting to get
cumbersome that it's not a maintained tool anymore. Add a configuration
for easy builds and try to get rid of the tool. It's a shame that it
works great when it decides to.